fix: move client config and info to client folder
