+++ data/fafb.yml
@@ -13,8 +13,8 @@ graph_config:
 backend_client:
   TYPE: "bigtable"
   CONFIG:
-    ADMIN: y
-    READ_ONLY: n
+    ADMIN: true
+    READ_ONLY: false
 
 ingest_config:
   AGGLOMERATION: "gs://ranl-scratch/190410_FAFB_v02_ws_size_threshold_200/agg"
