+++ pychunkedgraph/graph/chunkedgraph.py
@@ -11,11 +11,11 @@ from . import operation
 from . import attributes
 from . import exceptions
 from .client import base
-from .client.utils import get_default_client_info
+from .client import BackendClientInfo
+from .client import get_default_client_info
 from .client.bigtable import BigTableClient
 from .cache import CacheService
 from .meta import ChunkedGraphMeta
-from .meta import BackendClientInfo
 from .utils import basetypes
 from .utils import id_helpers
 from .utils import generic as misc_utils
