+++ pychunkedgraph/graph/client/__init__.py
@@ -16,3 +16,27 @@ making it possible to use any unique features these solutions may provide.
 
 Please see `base.py` for more details.
 """
+
+from collections import namedtuple
+
+
+_backend_clientinfo_fields = ("TYPE", "CONFIG")
+_backend_clientinfo_defaults = (None, None)
+BackendClientInfo = namedtuple(
+    "BackendClientInfo",
+    _backend_clientinfo_fields,
+    defaults=_backend_clientinfo_defaults,
+)
+
+
+def get_default_client_info():
+    """
+    Load client from env variables.
+    """
+
+    # TODO make dynamic after multiple platform support is added
+    from .bigtable import get_client_info as get_bigtable_client_info
+
+    return BackendClientInfo(
+        CONFIG=get_bigtable_client_info(admin=True, read_only=False)
+    )
