+++ pychunkedgraph/graph/meta.py
@@ -17,35 +17,6 @@ DataSource = namedtuple(
 )
 
 
-_bigtableconfig_fields = (
-    "PROJECT",
-    "INSTANCE",
-    "ADMIN",
-    "READ_ONLY",
-    "CREDENTIALS",
-)
-_bigtableconfig_defaults = (
-    "neuromancer-seung-import",
-    "pychunkedgraph",
-    True,
-    False,
-    None,
-)
-# TODO move this to bigtable client folder
-BigTableConfig = namedtuple(
-    "BigTableConfig", _bigtableconfig_fields, defaults=_bigtableconfig_defaults
-)
-
-
-_backend_clientinfo_fields = ("TYPE", "CONFIG")
-_backend_clientinfo_defaults = ("bigtable", BigTableConfig())
-BackendClientInfo = namedtuple(
-    "BackendClientInfo",
-    _backend_clientinfo_fields,
-    defaults=_backend_clientinfo_defaults,
-)
-
-
 _graphconfig_fields = (
     "ID",  # ID_PREFIX and ID are together used when creating the graph
     "ID_PREFIX",
